requirements.txt
absl-py==2.1.0
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2024.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
Django==4.2.11
django-appconf==1.0.6
django-crispy-forms==2.2
django-filter==24.2
django-select2==8.1.2
django-tables2==2.7.0
django-taggit==5.0.1
django-viewflow==2.1.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
executing==2.0.1
fastjsonschema==2.20.0
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.6
inflection==0.5.1
ipykernel==6.29.5
ipython==8.26.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.2
tenacity==8.4.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xgboost==2.1.0
